+++ src/py21cmfast/src/Constants.h
@@ -11,8 +11,7 @@
 #ifndef _CONSTANTS_H
 #define _CONSTANTS_H
 
-//TODO: sort out these constants into groups and make separate files
-#include "Globals.h"
+#include "InputParameters.h"
 
 // ----------------------------------------------------------------------------------------- //
