+++ src/py21cmfast/src/HaloBox.c
@@ -9,7 +9,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
 #include "cosmology.h"
