+++ src/py21cmfast/src/IonisationBox.c
@@ -10,7 +10,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
 #include "cosmology.h"
