+++ src/py21cmfast/src/LuminosityFunction.c
@@ -11,7 +11,6 @@
 #include "exceptions.h"
 #include "logger.h"
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
