+++ src/py21cmfast/src/Stochasticity.c
@@ -9,7 +9,6 @@
 #include "exceptions.h"
 #include "logger.h"
 #include "Constants.h"
-#include "Globals.h"
 #include "indexing.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
