+++ src/py21cmfast/src/dft.c
@@ -8,7 +8,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
 #include "cosmology.h"
