+++ src/py21cmfast/src/filtering.c
@@ -12,7 +12,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "indexing.h"
