+++ src/py21cmfast/src/heating_helper_progs.c
@@ -11,7 +11,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "indexing.h"
