+++ src/py21cmfast/src/hmf.c
@@ -10,7 +10,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "cosmology.h"
