+++ src/py21cmfast/src/thermochem.c
@@ -7,7 +7,6 @@
 #include "logger.h"
 
 #include "Constants.h"
-#include "Globals.h"
 #include "InputParameters.h"
 #include "OutputStructs.h"
 #include "thermochem.h"
